+++ test/fixtures/warn.js
@@ -1,0 +1,2 @@
+/*eslint no-unused-vars: 1*/
+var foo = this;